+++ sdk/program/src/lamports.rs
@@ -1,0 +1,22 @@
+use crate::instruction::InstructionError;
+use thiserror::Error;
+
+#[derive(Debug, Error)]
+pub enum LamportsError {
+    /// arithmetic underflowed
+    #[error("Arithmetic underflowed")]
+    ArithmeticUnderflow,
+
+    /// arithmetic overflowed
+    #[error("Arithmetic overflowed")]
+    ArithmeticOverflow,
+}
+
+impl From<LamportsError> for InstructionError {
+    fn from(error: LamportsError) -> Self {
+        match error {
+            LamportsError::ArithmeticOverflow => InstructionError::ArithmeticOverflow,
+            LamportsError::ArithmeticUnderflow => InstructionError::ArithmeticOverflow,
+        }
+    }
+}
